Se corrige error en cierre mes. Monto pedido estaba como int y daba error. Se pasa a Double

+++ Negocio/Controladora.cs
@@ -1463,7 +1463,7 @@ namespace Negocio
                     cedula = dsCobranzas.Tables["cobranzas"].Rows[i][2].ToString();
                     tasa = Convert.ToDouble(dsCobranzas.Tables["cobranzas"].Rows[i][3].ToString());
                     Wiva = Convert.ToDouble(dsCobranzas.Tables["cobranzas"].Rows[i][4].ToString()); // Porcentaje ivaç
-                    montoPedido = Convert.ToInt32(dsCobranzas.Tables["cobranzas"].Rows[i][5].ToString());
+                    montoPedido = Convert.ToDouble(dsCobranzas.Tables["cobranzas"].Rows[i][5].ToString());
                     CuotasVan = Convert.ToInt32(dsCobranzas.Tables["cobranzas"].Rows[i][7].ToString()) + 1;
                     cantidadCuotas = Convert.ToInt32(dsCobranzas.Tables["cobranzas"].Rows[i][6].ToString());
                     importeCuota = Convert.ToDouble(dsCobranzas.Tables["cobranzas"].Rows[i][8].ToString());
